Removed execute permission on files
Best that I can determine, execute permission is not required
and should not be set on any of the WordPress files.

The execute permission is a potential security risk.

The variability in the setting of the execute permission causes merge
and difference problems.

The execute permission is often unintentionally turned on when development
is done on a Microsoft computer.  MS has very different handling and use
of the execute permission.

For these reasons, we will attempt to turn off the execute bit on future
commits.