zoomit-wordpress 0.4.2
Not available as a plugin from WordPress.
Found it at github.

git://github.com/openzoom/zoomit-wordpress.git
git://github.com/openzoom/zoomit-wordpress.git

Note that this plugin sends documents to an external server which
provides a URL to the zoomed version.  Need to check out this site
and their policies.